fix test concurrency
