--- test/test.js
@@ -1,123 +1,0 @@
-const test = require('ava');
-const cfntest = require('@cfn-modules/test');
-
-test('access-cloud-front-access-log-write', async t => {
-  const stackName = cfntest.stackName();
-  try {
-    t.log(await cfntest.createStack(`${__dirname}/access-cloud-front-access-log-write.yml`, stackName, {}));
-    // what could we test here?
-  } finally {
-    t.log(await cfntest.deleteStack(stackName));
-    t.pass();
-  }
-});
-
-test('access-cloud-front-read', async t => {
-  const stackName = cfntest.stackName();
-  try {
-    t.log(await cfntest.createStack(`${__dirname}/access-cloud-front-read.yml`, stackName, {}));
-    // what could we test here?
-  } finally {
-    t.log(await cfntest.deleteStack(stackName));
-    t.pass();
-  }
-});
-
-test('access-cloud-trail-write', async t => {
-  const stackName = cfntest.stackName();
-  try {
-    t.log(await cfntest.createStack(`${__dirname}/access-cloud-trail-write.yml`, stackName, {}));
-    // what could we test here?
-  } finally {
-    t.log(await cfntest.deleteStack(stackName));
-    t.pass();
-  }
-});
-
-test('access-config-write', async t => {
-  const stackName = cfntest.stackName();
-  try {
-    t.log(await cfntest.createStack(`${__dirname}/access-config-write.yml`, stackName, {}));
-    // what could we test here?
-  } finally {
-    t.log(await cfntest.deleteStack(stackName));
-    t.pass();
-  }
-});
-
-test('access-elb-access-log-write', async t => {
-  const stackName = cfntest.stackName();
-  try {
-    t.log(await cfntest.createStack(`${__dirname}/access-elb-access-log-write.yml`, stackName, {}));
-    // what could we test here?
-  } finally {
-    t.log(await cfntest.deleteStack(stackName));
-    t.pass();
-  }
-});
-
-test('access-public-read', async t => {
-  const stackName = cfntest.stackName();
-  try {
-    t.log(await cfntest.createStack(`${__dirname}/access-public-read.yml`, stackName, {}));
-    // what could we test here?
-  } finally {
-    t.log(await cfntest.deleteStack(stackName));
-    t.pass();
-  }
-});
-
-test('cors-allow-all', async t => {
-  const stackName = cfntest.stackName();
-  try {
-    t.log(await cfntest.createStack(`${__dirname}/cors-allow-all.yml`, stackName, {}));
-    // what could we test here?
-  } finally {
-    t.log(await cfntest.deleteStack(stackName));
-    t.pass();
-  }
-});
-
-test('defaults', async t => {
-  const stackName = cfntest.stackName();
-  try {
-    t.log(await cfntest.createStack(`${__dirname}/defaults.yml`, stackName, {}));
-    // what could we test here?
-  } finally {
-    t.log(await cfntest.deleteStack(stackName));
-    t.pass();
-  }
-});
-
-test('with-bucketname', async t => {
-  const stackName = cfntest.stackName();
-  try {
-    t.log(await cfntest.createStack(`${__dirname}/with-bucketname.yml`, stackName, {}));
-    // what could we test here?
-  } finally {
-    t.log(await cfntest.deleteStack(stackName));
-    t.pass(); 
-  }
-});
-
-test('with-kms', async t => {
-  const stackName = cfntest.stackName();
-  try {
-    t.log(await cfntest.createStack(`${__dirname}/with-kms.yml`, stackName, {}));
-    // what could we test here?
-  } finally {
-    t.log(await cfntest.deleteStack(stackName));
-    t.pass(); 
-  }
-});
-
-test('with-lambda', async t => {
-  const stackName = cfntest.stackName();
-  try {
-    t.log(await cfntest.createStack(`${__dirname}/with-lambda.yml`, stackName, {}));
-    // what could we test here?
-  } finally {
-    t.log(await cfntest.deleteStack(stackName));
-    t.pass(); 
-  }
-});
